tools/requirements.txt
django-import-export
numpy
pandas
plotly